third_party/blink/web_tests/wpt_internal/bluetooth/requestDevice/name-missing-device-from-name-wrong-filter.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script>
'use strict';
bluetooth_test(t => {
  return setBluetoothFakeAdapter('NoNameHeartRateAdapter')
      .then(
          () => promise_rejects_dom(
              t, 'NotFoundError',
              requestDeviceWithTrustedClick(
                  {filters: [{name: 'a', services: ['heart_rate']}]})));
}, 'An unnamed device can not be obtained by name filter.');
</script>
